requirements.txt
datasets
torch
scikit-learn
Pillow
numpy
transformers